pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cell-size"
version = "0.1.0"
description = "Batch cell segmentation and size estimation using Cellpose-SAM with Hydra configuration"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "hydra-core>=1.3",
    "numpy",
    "tifffile",
    "imagecodecs",
    "opencv-python-headless",
    "matplotlib",
    "natsort",
    "pandas",
    "scikit-image",
    "torch",
    "torchvision",
    "scikit-learn",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]
logging = [
    "wandb",
]
demo = [
    "gradio",
]

[project.scripts]
cell-size = "cell_size.main:main"
cell-size-train = "cell_size.train_main:main"
cell-size-classify = "cell_size.classify_main:main"
cell-size-demo = "cell_size.demo_launcher:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cell_size = ["configs/**/*.yaml"]

[tool.ruff]
line-length = 100
target-version = "py310"
